V1 Added
As for now, I'll leave it with only one level because I have no time to submit. I only miss out on game instructions like "press WASD to move" or "press CTRL to slide", those stuff. besides that, its only a matter of level design. everything in the game is somewhere of 90% completed. This is made entirely on Godot 4.4 and took over 60 hours to make. I hope to get this improved more and more because I see a vision in this type of game. If anyone think he/she can help, I'm more than open to collaborate and turn this prototype into a finished, packed game.\